// lib/prisma/enums.ts
export const ROLE = {
  SYSTEM_OWNER: "SYSTEM_OWNER",
  SYSTEM_ADMIN: "SYSTEM_ADMIN",
  SYSTEM_USER: "SYSTEM_USER",
  USER: "USER",
} as const;

export type ROLE = (typeof ROLE)[keyof typeof ROLE];